feat: refactor showSnackbar to be included on main component
